# ajnaapi/openapi.yaml
openapi: 3.0.1
info:
  title: API-AJNA
  description: API para acesso de outras aplicações a dados do AJNA
  version: "1.0b"
  termsOfService: "https://ajna.labin08.rf08.srf/"
  contact:
    url: "https://ajna.labin08.rf08.srf/"
    name: "Receita Federal"
    email: "ivan.brasilico@rfb.gov.br"
  license:
    name: Receita Federal C-level site
    url: "https://ajna.labin08.rf08.srf/site"
...
servers:
  - url: /ajnaapi/api
...
paths:
...
  /ficha/{id}:
    get:
      summary: Retorna conteúdo da Ficha de Controle de Carga
      tags:
        - bhadrasana
      parameters:
        - name: id
          in: path
          description: id da Ficha
          required: true
          schema:
            type: string
      responses:
        200:
          description: OVR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OVR'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /fichas:
    post:
      summary: Retorna lista de Fichas de Controle de Carga
      tags:
        - bhadrasana
      requestBody:
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/OVRFiltro'
      responses:
        200:
          description: OVR
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/OVR'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /rvf/{id}:
    get:
      summary: Retorna conteúdo da Verificação Física
      tags:
        - bhadrasana
      parameters:
        - name: id
          in: path
          description: id da Ficha
          required: true
          schema:
            type: string
      responses:
        200:
          description: RVF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RVF'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /rvfs/{ovr_id}:
    get:
      summary: Retorna conteúdos da Verificações Físicas da OVR
      tags:
        - bhadrasana
      parameters:
        - name: ovr_id
          in: path
          description: id da Ficha
          required: true
          schema:
            type: string
      responses:
        200:
          description: RVF
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/RVF'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /rvfs:
    post:
      summary: Retorna lista de Fichas de Controle de Carga
      tags:
        - bhadrasana
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RVFFiltro'
      responses:
        200:
          description: RVF
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/RVF'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /imagens_rvf/{rvf_id}:
    get:
      summary: Retorna lista de imagens da RVF
      tags:
        - bhadrasana
      parameters:
        - name: rvf_id
          in: path
          description: id da RVF
          required: true
          schema:
            type: string
      responses:
        200:
          description: ImagemRVF
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ImagemRVF'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /tg/{id}:
    get:
      summary: Retorna conteúdo do TG
      tags:
        - bhadrasana
      parameters:
        - name: id
          in: path
          description: id do TG
          required: true
          schema:
            type: string
      responses:
        200:
          description: TGOVR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TGOVR'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /tgs/{ovr_id}:
    get:
      summary: Retorna lista de TGs do OVR
      tags:
        - bhadrasana
      parameters:
        - name: ovr_id
          in: path
          description: id da Ficha
          required: true
          schema:
            type: string
      responses:
        200:
          description: TGOVR
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/TGOVR'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /tgs:
    post:
      summary: Retorna lista de TGs conforme o Filtro
      tags:
        - bhadrasana
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TGFiltro'
      responses:
        200:
          description: TGOVR
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/TGOVR'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /itemtg/{id}:
    get:
      summary: Retorna conteúdo do item do TG
      tags:
        - bhadrasana
      parameters:
        - name: id
          in: path
          description: id do itemtg
          required: true
          schema:
            type: string
      responses:
        200:
          description: ItemTG
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemTG'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /itenstg/{tg_id}:
    get:
      summary: Retorna lista itens do TG
      tags:
        - bhadrasana
      parameters:
        - name: tg_id
          in: path
          description: id do tg
          required: true
          schema:
            type: string
      responses:
        200:
          description: ItemTG
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ItemTG'
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /minhas_fichas/{cpf}:
    get:
      summary: "Retorna lista de fichas atribuídas ao CPF para visualização"
      tags:
        - bhadrasana
      parameters:
        - name: cpf
          in: path
          description: CPF a consultar
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Lista de dicionários contendo campos selecionados da OVR"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties:
                    type: string
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /get_cpf_telegram/{telegram_user}:
    get:
      summary: "Retorna cpf do usuario se username telegram estiver cadastrado"
      tags:
        - bhadrasana
      parameters:
        - name: telegram_user
          in: path
          description: Username telegram
          required: true
          schema:
            type: string
      responses:
        200:
          description: CPF correspondente ao telegram
          content:
            application/json:
              schema:
                type: object
                properties:
                  cpf:
                    type: string
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /consulta_conteiner:
    post:
      summary: "Retorna resumo de informações para o conteiner {numero}: OVRs, RVFs, CEs, DUEs, Eventos"
      tags:
        - bhadrasana
      requestBody:
        description: Numero do container, data inicio e datafim
        content:
          application/json:
            schema:
              type: object
              properties:
                numerolote:
                  type: string
                datainicio:
                  type: string
                datafim:
                  type: string
      responses:
        200:
          description: "Lista de dicionários contendo campos selecionados dos sistemas"
          content:
            application/json:
              schema:
                type: object
                properties:
                  rvfs:
                    type: object
                    properties:
                      campo:
                        type: string
                  ovrs:
                    type: object
                    properties:
                      campo:
                        type: string
                  infoces:
                    type: object
                    properties:
                      campo:
                        type: string
                  dues:
                    type: object
                    properties:
                      campo:
                        type: string
                  eventos:
                    type: object
                    properties:
                      campo:
                        type: string
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
  /consulta_empresa:
    post:
      summary: "Retorna resumo de informações para a empresa {cnpj}: OVRs, CEs, DUEs, SAT/LABs"
      tags:
        - bhadrasana
      requestBody:
        description: CNPJ da Empresa, data inicio e data fim
        content:
          application/json:
            schema:
              type: object
              properties:
                cnpj:
                  type: string
                datainicio:
                  type: string
                datafim:
                  type: string
      responses:
        200:
          description: "Lista de dicionários contendo campos selecionados dos sistemas"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ovrs:
                    type: object
                    properties:
                      campo:
                        type: string
                  infoces:
                    type: object
                    properties:
                      campo:
                        type: string
                  dues:
                    type: object
                    properties:
                      campo:
                        type: string
                  sats:
                    type: object
                    properties:
                      campo:
                        type: string
        400:
          description: Erro
          content: {}
        404:
          description: Não encontrado para filtro atual.
          content: {}
      security:
        - jwt: ['secret']
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: authentication.decode_token
  schemas:
...
    OVR:
      type: object
      properties:
        id:
          type: integer
        ano:
          type: string
        cnpj_fiscalizado:
          type: string
        create_date:
          type: string
        dataentrada:
          type: string
        datahora:
          type: string
        fase:
          type: integer
        flags:
          type: array
          items:
            type: string
        historico:
          type: array
          items:
            type: object
            $ref: '#components/schemas/EventoOVR'
        last_modified:
          type: string
        numero:
          type: string
        numeroCEmercante:
          type: string
        numerodeclaracao:
          type: string
        observacoes:
          type: string
        recinto_id:
          type: integer
        responsavel_cpf:
          type: string
        setor_id:
          type: string
        tgs:
          type: array
          items:
            type: object
            $ref: '#components/schemas/TGOVR'
        tipoevento_id:
          type: integer
        tipooperacao:
          type: integer
        user_name:
          type: string
    OVRFiltro:
      type: object
      properties:
        cnpj_fiscalizado:
          type: string
        datahora:
          type: string
        fase:
          type: integer
        numero:
          type: string
        numeroCEmercante:
          type: string
        numerodeclaracao:
          type: string
        recinto_id:
          type: integer
        responsavel_cpf:
          type: string
        setor_id:
          type: string
        tipoevento_id:
          type: integer
        tipooperacao:
          type: integer
        user_name:
          type: string
    RVF:
      type: object
      properties:
        id:
          type: integer
        create_date:
          type: string
        datahora:
          type: string
        descricao:
          type: string
        numeroCEmercante:
          type: string
        numerolote:
          type: string
        ovr_id:
          type: integer
        peso:
          type: number
        user_name:
          type: string
        volume:
          type: number
        last_modified:
          type: string
        imagens:
          type: array
          items:
            type: object
            $ref: '#components/schemas/ImagemRVF'
        infracoesencontradas:
          type: array
          items:
            type: string
        lacresverificados:
          type: array
          items:
            type: string
        marcasencontradas:
          type: array
          items:
            type: string
    RVFFiltro:
      type: object
      properties:
        datahora:
          type: string
        descricao:
          type: string
        numeroCEmercante:
          type: string
        numerolote:
          type: string
        user_name:
          type: string
    ImagemRVF:
      type: object
      properties:
        id:
          type: integer
        create_date:
          type: string
        descricao:
          type: string
        imagem:
          type: string
        itemtg_id:
          type: integer
        marca:
          type: string
        ordem:
          type: integer
        rvf_id:
          type: integer
        tg_id:
          type: integer
        user_name:
          type: string
...
    TGOVR:
      type: object
      properties:
        id:
          type: integer
        afrfb:
          type: string
        create_date:
          type: string
        descricao:
          type: string
        identificacao:
          type: string
        itenstg:
          type: array
          items:
            type: object
            $ref: '#components/schemas/ItemTG'
        marcas:
          type: array
          items:
            type: string
        numerolote:
          type: string
        numerotg:
          type: string
        observacoes:
          type: string
        ovr_id:
          type: integer
        qtde:
          type: number
        tipomercadoria_id:
          type: integer
        unidadedemedida:
          type: integer
        user_name:
          type: string
        valor:
          type: number
    TGFiltro:
      type: object
      properties:
        afrfb:
          type: string
        descricao:
          type: string
        identificacao:
          type: string
        numerolote:
          type: string
        numerotg:
          type: string
        observacoes:
          type: string
        tipomercadoria_id:
          type: integer
    ItemTG:
      type: object
      properties:
        id:
          type: integer
        create_date:
          type: string
        descricao:
          type: string
        marca:
          type: string
        ncm:
          type: string
        numero:
          type: integer
        qtde:
          type: number
        unidadedemedida:
          type: integer
        user_name:
          type: string
        valor:
          type: number